wtp/webrtc/RTCRtpTransceiver-setCodecPreferences.html

<!doctype html>
<meta charset="utf-8">
<title>RTCRtpTransceiver.prototype.setCodecPreferences</title>
<script src="/resources/webrtc-ie-shim.js"></script><script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  'use strict';
  
  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html
  
  /*
    5.4.  RTCRtpTransceiver Interface
      interface RTCRtpTransceiver {
        ...
        void setCodecPreferences(sequence<RTCRtpCodecCapability> codecs);
      };
       setCodecPreferences
        - Setting codecs to an empty sequence resets codec preferences to any
          default value.
         - The codecs sequence passed into setCodecPreferences can only contain
          codecs that are returned by RTCRtpSender.getCapabilities(kind) or
          RTCRtpReceiver.getCapabilities(kind), where kind is the kind of the
          RTCRtpTransceiver on which the method is called. Additionally, the
          RTCRtpCodecParameters dictionary members cannot be modified. If
          codecs does not fulfill these requirements, the user agent MUST throw
          an InvalidAccessError.
   */
  
  var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"]) _i["return"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError("Invalid attempt to destructure non-iterable instance"); } }; }();
  
  function _toConsumableArray(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = Array(arr.length); i < arr.length; i++) { arr2[i] = arr[i]; } return arr2; } else { return Array.from(arr); } }
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var capabilities = RTCRtpSender.getCapabilities('audio');
    transceiver.setCodecPreferences(capabilities.codecs);
  }, 'setCodecPreferences() on audio transceiver with codecs returned from RTCRtpSender.getCapabilities(\'audio\') should succeed');
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('video');
    var capabilities = RTCRtpReceiver.getCapabilities('video');
    transceiver.setCodecPreferences(capabilities.codecs);
  }, 'setCodecPreferences() on video transceiver with codecs returned from RTCRtpReceiver.getCapabilities(\'video\') should succeed');
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var capabilities1 = RTCRtpSender.getCapabilities('audio');
    var capabilities2 = RTCRtpReceiver.getCapabilities('audio');
    transceiver.setCodecPreferences([].concat(_toConsumableArray(capabilities1.codecs), _toConsumableArray(capabilities2.codecs)));
  }, 'setCodecPreferences() with both sender receiver codecs combined should succeed');
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    transceiver.setCodecPreferences([]);
  }, 'setCodecPreferences([]) should succeed');
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var capabilities = RTCRtpSender.getCapabilities('audio');
    var codecs = capabilities.codecs;
  
  
    if (codecs.length >= 2) {
      var tmp = codecs[0];
      codecs[0] = codecs[1];
      codecs[1] = tmp;
    }
  
    transceiver.setCodecPreferences(codecs);
  }, 'setCodecPreferences() with reordered codecs should succeed');
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var capabilities = RTCRtpSender.getCapabilities('video');
    assert_throws(function () {
      return transceiver.setCodecPreferences(capabilities.codecs);
    });
  }, 'setCodecPreferences() on audio transceiver with codecs returned from getCapabilities(\'video\') should throw InvalidAccessError');
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var codecs = [{
      mimeType: 'audio/piepiper',
      clockRate: 2000,
      channels: 2,
      sdpFmtpLine: 'a=fmtp:98 0-15'
    }];
  
    assert_throws(function () {
      return transceiver.setCodecPreferences(codecs);
    });
  }, 'setCodecPreferences() with user defined codec should throw InvalidAccessError');
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var capabilities = RTCRtpSender.getCapabilities('audio');
    var codecs = [].concat(_toConsumableArray(capabilities.codecs), [{
      mimeType: 'audio/piepiper',
      clockRate: 2000,
      channels: 2,
      sdpFmtpLine: 'a=fmtp:98 0-15'
    }]);
  
    assert_throws(function () {
      return transceiver.setCodecPreferences(codecs);
    });
  }, 'setCodecPreferences() with user defined codec together with codecs returned from getCapabilities() should throw InvalidAccessError');
  
  test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var capabilities = RTCRtpSender.getCapabilities('audio');
  
    var codecs = capabilities.codecs;
  
    assert_greater_than(codecs.length, 0, 'Expect at least one codec available');
  
    var _codecs = _slicedToArray(codecs, 1),
        codec = _codecs[0];
  
    var _codec$channels = codec.channels,
        channels = _codec$channels === undefined ? 2 : _codec$channels;
  
    codec.channels = channels + 1;
  
    assert_throws(function () {
      return transceiver.setCodecPreferences(codecs);
    });
  }, 'setCodecPreferences() with modified codecs returned from getCapabilities() should throw InvalidAccessError');

 </script>
